requirements.txt
Flask
pandas
numpy
scikit-learn
nltk
pymorphy2
requests
gensim
tqdm
gunicorn
python-dotenv
surprise
implicit